# docs/openapi.json
{
  "components" : {
    "schemas" : {
      "Actual" : {
        "properties" : {
          "breakdown" : {
            "$ref" : "#/components/schemas/FareBreakdownResponse"
          },
          "distanceMeters" : {
            "format" : "int32",
            "type" : "integer"
          },
          "distanceSource" : {
            "enum" : [ "TRACKED", "ESTIMATED" ],
            "type" : "string"
          },
          "durationSeconds" : {
            "format" : "int32",
            "type" : "integer"
          },
          "fare" : {
            "$ref" : "#/components/schemas/Money"
          }
        },
        "required" : [ "breakdown", "distanceMeters", "distanceSource", "durationSeconds", "fare" ],
        "type" : "object"
      },
...
      "AdminRideAnalysis" : {
        "properties" : {
          "failureCode" : {
            "type" : [ "string", "null" ]
          },
          "status" : {
            "enum" : [ "PENDING", "COMPLETED", "FAILED", "UNAVAILABLE" ],
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "required" : [ "status", "updatedAt" ],
        "type" : "object"
      },
      "AdminRideDetailResponse" : {
        "properties" : {
          "analysis" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/AdminRideAnalysis"
            }, {
              "type" : "null"
            } ]
          },
          "offers" : {
            "items" : {
              "$ref" : "#/components/schemas/AdminRideOffer"
            },
            "type" : "array"
          },
          "passenger" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/AdminRidePassenger"
            }, {
              "type" : "null"
            } ]
          },
          "ride" : {
            "$ref" : "#/components/schemas/RideResponse"
          },
          "timeline" : {
            "items" : {
              "$ref" : "#/components/schemas/RideTimelineEntryResponse"
            },
            "type" : "array"
          }
        },
        "required" : [ "offers", "ride", "timeline" ],
        "type" : "object"
      },
      "AdminRideOffer" : {
        "properties" : {
          "distanceMeters" : {
            "format" : "int32",
            "type" : "integer"
          },
          "driverId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "expiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "offeredAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "respondedAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "round" : {
            "format" : "int32",
            "type" : "integer"
          },
          "status" : {
            "enum" : [ "PENDING", "ACCEPTED", "REJECTED", "EXPIRED", "CANCELLED" ],
            "type" : "string"
          }
        },
        "required" : [ "distanceMeters", "driverId", "expiresAt", "offeredAt", "round", "status" ],
        "type" : "object"
      },
      "AdminRidePassenger" : {
        "properties" : {
          "email" : {
            "type" : "string"
          },
          "fullName" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "required" : [ "email", "fullName", "id" ],
        "type" : "object"
      },
...
      "BookRideRequest" : {
        "properties" : {
          "dropoff" : {
            "$ref" : "#/components/schemas/PlaceRequest"
          },
          "paymentMethod" : {
            "enum" : [ "CASH", "CARD" ],
            "type" : "string"
          },
          "pickup" : {
            "$ref" : "#/components/schemas/PlaceRequest"
          },
          "quoteId" : {
            "maxLength" : 4096,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "dropoff", "paymentMethod", "pickup", "quoteId" ],
        "type" : "object"
      },
      "CancelRideRequest" : {
        "properties" : {
          "reason" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : [ "string", "null" ]
          }
        },
        "type" : "object"
      },
...
      "Money" : {
        "properties" : {
          "amount" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          }
        },
        "required" : [ "amount", "currency" ],
        "type" : "object"
      },
...
      "PageResponseRideSummaryResponse" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/RideSummaryResponse"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sort" : {
            "type" : "string"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "content", "page", "size", "sort", "totalElements", "totalPages" ],
        "type" : "object"
      },
...
      "PlaceRequest" : {
        "properties" : {
          "address" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "point" : {
            "$ref" : "#/components/schemas/GeoPoint"
          }
        },
        "required" : [ "address", "point" ],
        "type" : "object"
      },
...
      "RateRideRequest" : {
        "properties" : {
          "comment" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : [ "string", "null" ]
          },
          "score" : {
            "format" : "int32",
            "maximum" : 5,
            "minimum" : 1,
            "type" : "integer"
          }
        },
        "required" : [ "score" ],
        "type" : "object"
      },
      "RatingResponse" : {
        "properties" : {
          "comment" : {
            "type" : [ "string", "null" ]
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "rideId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "score" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "required" : [ "createdAt", "id", "rideId", "score" ],
        "type" : "object"
      },
...
      "RideResponse" : {
        "properties" : {
          "actual" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/Actual"
            }, {
              "type" : "null"
            } ]
          },
          "cancellation" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/Cancellation"
            }, {
              "type" : "null"
            } ]
          },
          "driver" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/DriverInfo"
            }, {
              "type" : "null"
            } ]
          },
          "dropoff" : {
            "$ref" : "#/components/schemas/RidePlace"
          },
          "estimate" : {
            "$ref" : "#/components/schemas/Estimate"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "matching" : {
            "$ref" : "#/components/schemas/Matching"
          },
          "passenger" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/PassengerInfo"
            }, {
              "type" : "null"
            } ]
          },
          "payment" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/PaymentInfo"
            }, {
              "type" : "null"
            } ]
          },
          "paymentMethod" : {
            "enum" : [ "CASH", "CARD" ],
            "type" : "string"
          },
          "pickup" : {
            "$ref" : "#/components/schemas/RidePlace"
          },
          "status" : {
            "enum" : [ "REQUESTED", "MATCHING", "DRIVER_ASSIGNED", "DRIVER_ARRIVING", "DRIVER_ARRIVED", "IN_PROGRESS", "COMPLETED", "CANCELLED", "EXPIRED" ],
            "type" : "string"
          },
          "timestamps" : {
            "$ref" : "#/components/schemas/Timestamps"
          },
          "vehicleCategory" : {
            "enum" : [ "ECONOMY", "COMFORT", "XL" ],
            "type" : "string"
          },
          "version" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "dropoff", "estimate", "id", "matching", "paymentMethod", "pickup", "status", "timestamps", "vehicleCategory", "version" ],
        "type" : "object"
      },
      "RideSummaryResponse" : {
        "properties" : {
          "completedAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "dropoffAddress" : {
            "type" : "string"
          },
          "fare" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/Money"
            }, {
              "type" : "null"
            } ]
          },
          "fareIsFinal" : {
            "type" : "boolean"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "pickupAddress" : {
            "type" : "string"
          },
          "requestedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "enum" : [ "REQUESTED", "MATCHING", "DRIVER_ASSIGNED", "DRIVER_ARRIVING", "DRIVER_ARRIVED", "IN_PROGRESS", "COMPLETED", "CANCELLED", "EXPIRED" ],
            "type" : "string"
          },
          "vehicleCategory" : {
            "enum" : [ "ECONOMY", "COMFORT", "XL" ],
            "type" : "string"
          }
        },
        "required" : [ "dropoffAddress", "fareIsFinal", "id", "pickupAddress", "requestedAt", "status", "vehicleCategory" ],
        "type" : "object"
      },
      "RideTimelineEntryResponse" : {
        "properties" : {
          "actor" : {
            "enum" : [ "PASSENGER", "DRIVER", "SYSTEM", "ADMIN" ],
            "type" : "string"
          },
          "from" : {
            "enum" : [ "REQUESTED", "MATCHING", "DRIVER_ASSIGNED", "DRIVER_ARRIVING", "DRIVER_ARRIVED", "IN_PROGRESS", "COMPLETED", "CANCELLED", "EXPIRED" ],
            "type" : [ "string", "null" ]
          },
          "occurredAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "reason" : {
            "type" : [ "string", "null" ]
          },
          "rideVersion" : {
            "format" : "int64",
            "type" : "integer"
          },
          "to" : {
            "enum" : [ "REQUESTED", "MATCHING", "DRIVER_ASSIGNED", "DRIVER_ARRIVING", "DRIVER_ARRIVED", "IN_PROGRESS", "COMPLETED", "CANCELLED", "EXPIRED" ],
            "type" : "string"
          }
        },
        "required" : [ "actor", "occurredAt", "rideVersion", "to" ],
        "type" : "object"
      },
      "RideTrackingResponse" : {
        "properties" : {
          "driverLocation" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/DriverLocation"
            }, {
              "type" : "null"
            } ]
          },
          "eta" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/EtaResponse"
            }, {
              "type" : "null"
            } ]
          },
          "rideId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "stale" : {
            "type" : "boolean"
          },
          "status" : {
            "enum" : [ "REQUESTED", "MATCHING", "DRIVER_ASSIGNED", "DRIVER_ARRIVING", "DRIVER_ARRIVED", "IN_PROGRESS", "COMPLETED", "CANCELLED", "EXPIRED" ],
            "type" : "string"
          }
        },
        "required" : [ "rideId", "stale", "status" ],
        "type" : "object"
      },
...
      "TripAnalysisResponse" : {
        "properties" : {
          "failureCode" : {
            "enum" : [ "TIMEOUT", "PROVIDER_ERROR", "RATE_LIMITED", "INVALID_RESPONSE", "REFUSED", "BUSY", "UNAVAILABLE" ],
            "type" : [ "string", "null" ]
          },
          "insights" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/TripInsights"
            }, {
              "type" : "null"
            } ]
          },
          "model" : {
            "type" : [ "string", "null" ]
          },
          "observations" : {
            "items" : {
              "$ref" : "#/components/schemas/TripObservationResponse"
            },
            "type" : "array"
          },
          "promptVersion" : {
            "type" : "string"
          },
          "provider" : {
            "type" : [ "string", "null" ]
          },
          "rideId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "status" : {
            "enum" : [ "PENDING", "COMPLETED", "FAILED", "UNAVAILABLE" ],
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "observations", "promptVersion", "rideId", "status" ],
        "type" : "object"
      },
...
      "TripQuestionRequest" : {
        "properties" : {
          "question" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "question" ],
        "type" : "object"
      },
      "TripQuestionResponse" : {
        "properties" : {
          "answer" : {
            "type" : [ "string", "null" ]
          },
          "answerable" : {
            "type" : [ "boolean", "null" ]
          },
          "askedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "factKeysUsed" : {
            "items" : {
              "type" : "string"
            },
            "type" : [ "array", "null" ]
          },
          "failureCode" : {
            "enum" : [ "TIMEOUT", "PROVIDER_ERROR", "RATE_LIMITED", "INVALID_RESPONSE", "REFUSED", "BUSY", "UNAVAILABLE" ],
            "type" : [ "string", "null" ]
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "question" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "COMPLETED", "FAILED", "UNAVAILABLE" ],
            "type" : "string"
          }
        },
        "required" : [ "askedAt", "id", "question", "status" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Real-time ride-hailing platform API. Conventions: docs/api.md",
    "title" : "RideFlow API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/admin/rides/{rideId}" : {
      "get" : {
        "operationId" : "detail",
        "parameters" : [ {
          "in" : "path",
          "name" : "rideId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminRideDetailResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "A ride with its passenger, status history, offers and AI analysis status",
        "tags" : [ "Admin: rides" ]
      }
    },
...
    "/api/rides" : {
      "get" : {
        "operationId" : "history",
        "parameters" : [ {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "enum" : [ "REQUESTED", "MATCHING", "DRIVER_ASSIGNED", "DRIVER_ARRIVING", "DRIVER_ARRIVED", "IN_PROGRESS", "COMPLETED", "CANCELLED", "EXPIRED" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "maximum" : 100,
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "default" : "requestedAt,desc",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseRideSummaryResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Trip history: rides booked (passenger) or driven (driver)",
        "tags" : [ "Rides" ]
      },
      "post" : {
        "operationId" : "book",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BookRideRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RideResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "summary" : "Book a ride from a fare quote; matching starts asynchronously",
        "tags" : [ "Rides" ]
      }
    },
    "/api/rides/active" : {
      "get" : {
        "operationId" : "active",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RideResponse"
                }
              }
            },
            "description" : "The ongoing ride"
          },
          "204" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RideResponse"
                }
              }
            },
            "description" : "No ongoing ride"
          }
        },
        "summary" : "The caller's ongoing ride, or 204 if none (used to restore state after reload)",
        "tags" : [ "Rides" ]
      }
    },
    "/api/rides/{rideId}" : {
      "get" : {
        "operationId" : "get",
        "parameters" : [ {
          "in" : "path",
          "name" : "rideId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RideResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Rides" ]
      }
    },
...
    "/api/rides/{rideId}/cancel" : {
      "post" : {
        "operationId" : "cancel",
        "parameters" : [ {
          "in" : "path",
          "name" : "rideId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CancelRideRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RideResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Passenger cancels; a driver before pickup releases the ride for re-dispatch",
        "tags" : [ "Rides" ]
      }
    },
...
    "/api/rides/{rideId}/rating" : {
      "post" : {
        "operationId" : "rate",
        "parameters" : [ {
          "in" : "path",
          "name" : "rideId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RateRideRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RatingResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "summary" : "Rate the other participant of a completed ride (once per participant)",
        "tags" : [ "Rides" ]
      }
    },
...
    "/api/rides/{rideId}/timeline" : {
      "get" : {
        "operationId" : "timeline",
        "parameters" : [ {
          "in" : "path",
          "name" : "rideId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/RideTimelineEntryResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Status history of the ride",
        "tags" : [ "Rides" ]
      }
    },
    "/api/rides/{rideId}/tracking" : {
      "get" : {
        "operationId" : "tracking",
        "parameters" : [ {
          "in" : "path",
          "name" : "rideId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RideTrackingResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Driver position and ETA snapshot; live updates then arrive over WebSocket",
        "tags" : [ "Rides" ]
      }
    },
    "/api/trips/{rideId}/ai-analysis" : {
      "get" : {
        "operationId" : "analysis",
        "parameters" : [ {
          "in" : "path",
          "name" : "rideId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TripAnalysisResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Computed observations, plus the AI analysis once it is ready (status PENDING until then)",
        "tags" : [ "Trip insights" ]
      }
    },
    "/api/trips/{rideId}/ai-analysis/questions" : {
      "get" : {
        "operationId" : "questions",
        "parameters" : [ {
          "in" : "path",
          "name" : "rideId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/TripQuestionResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Questions asked about this trip, oldest first",
        "tags" : [ "Trip insights" ]
      },
      "post" : {
        "operationId" : "ask",
        "parameters" : [ {
          "in" : "path",
          "name" : "rideId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TripQuestionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TripQuestionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Ask about this trip; answered from the trip's recorded facts only (503 AI_UNAVAILABLE if AI is down)",
        "tags" : [ "Trip insights" ]
      }
    },
    "/api/trips/{rideId}/ai-analysis/regenerate" : {
      "post" : {
        "operationId" : "regenerate",
        "parameters" : [ {
          "in" : "path",
          "name" : "rideId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "202" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TripAnalysisResponse"
                }
              }
            },
            "description" : "Accepted"
          }
        },
        "summary" : "Run a FAILED, UNAVAILABLE or abandoned analysis again (asynchronously)",
        "tags" : [ "Trip insights" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "url" : "/"
  } ]
}